Initial commit
Initial push of files after working on them for a little while, currently should support functionality for going through the motions of crafting if it was receiving the correct clipboard data. TransAltRegEx is the main file to run, while posGrabber will be used later on my home machine to get the coordinates of where those currencies are within the screen. Item and item_test were used to test clipboard data and see if it was being grabbed and searched correctly