Added actual coordinates and modified mouse clicks
Used the grabber to get the 1080p coordinates of transmutes, alterations, augments, regals, scours, exalts, and the showcase. Currently the script has only been tested crafting a jewel, other items will need some modification which I will add soonTM. Also adjusted the way the program handles moving the mouse and clicking because sometimes it would click too early and not grab the currency, or pick up the item randomly. Added a spot to pause the grabber so that I could properly write down the coordinates as well.